+++ src/main/java/org/icgc/argo/program_service/mappers/MapperConfig.java
@@ -1,22 +1,16 @@
-package org.icgc.argo.program_service.converters;
+package org.icgc.argo.program_service.mappers;
 
 import com.google.protobuf.MessageOrBuilder;
-import org.mapstruct.CollectionMappingStrategy;
-import org.mapstruct.InjectionStrategy;
-import org.mapstruct.MapperConfig;
-import org.mapstruct.Mapping;
-import org.mapstruct.MappingInheritanceStrategy;
-import org.mapstruct.NullValueCheckStrategy;
-import org.mapstruct.ReportingPolicy;
+import org.mapstruct.*;
 
-@MapperConfig(
+@org.mapstruct.MapperConfig(
   unmappedTargetPolicy = ReportingPolicy.ERROR,
   injectionStrategy = InjectionStrategy.CONSTRUCTOR,
   nullValueCheckStrategy = NullValueCheckStrategy.ALWAYS,
   collectionMappingStrategy = CollectionMappingStrategy.ADDER_PREFERRED,
   mappingInheritanceStrategy =  MappingInheritanceStrategy.AUTO_INHERIT_ALL_FROM_CONFIG
 )
-public interface ConverterConfig {
+public interface MapperConfig {
   @Mapping(target = "unknownFields", ignore = true)
   @Mapping(target = "allFields", ignore = true)
   MessageOrBuilder configForProtoMessageOrBuilder(Object o);
